doc-ingestor/requirements.txt
pandas
psycopg2
python-dotenv 